requirements.txt

faiss-cpu==1.7.4
Flask==3.0.0
logger==1.4
mtcnn==0.1.1
numpy==1.26.3
opencv-python==4.9.0.80
pandas==2.1.4
pillow==10.2.0
pytorch-lightning==2.1.3
pytorch-metric-learning==2.4.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
scikit-learn==1.3.2
scipy==1.11.4
tensorflow==2.15.0.post1
torch==2.1.2
torchmetrics==1.2.1
torchvision==0.16.2
tqdm==4.66.1
urllib3==2.1.0
Werkzeug==3.0.1
pytest==7.4.4
